split algorithm code into separate header and associated source files
