+++ memory-voting-prototype/.gitignore
@@ -1,1 +1,3 @@
 prototype
+
+*.o
