+++ memory-voting-prototype/makefile
@@ -4,12 +4,16 @@ LDFLAGS =
 
 LINTER=cppcheck --enable=all --inconclusive --max-ctu-depth=4 -j 4 --library=posix
 
+SOURCE=$(wildcard *.c)
+OBJECTS=$(SOURCE:%.c=%.o)
+
 .PHONY: all clean test
 
 all: prototype
 
 clean:
 	rm -f prototype
+	rm -f $(OBJECTS)
 
 check: all test lint
 
@@ -19,5 +23,8 @@ lint:
 test: all
 	time ./prototype
 
-prototype:
-	$(CC) $(CFLAGS) -o $@ prototype-voting.c
+%.o: %.c
+	$(CC) $(CFLAGS) -c $< -o $@
+
+prototype: memory_correction.o
+	$(CC) $(CFLAGS) -o $@ prototype-voting.c $<
